+++ packages/core/src/@types/index.ts
@@ -1,6 +1,7 @@
 import { z } from "zod/v4"
 import { OAuthAccessTokenErrorResponse, OAuthAuthorizationErrorResponse, OAuthEnvSchema } from "@/schemas.js"
 import type { SerializeOptions } from "@aura-stack/router/cookie"
+import type { JWTVerifyOptions, EncryptOptions, JWTDecryptOptions } from "@aura-stack/jose"
 import type { JWTPayload } from "@/jose.js"
 import type { BuiltInOAuthProvider } from "@/oauth/index.js"
 import type { LiteralUnion, Prettify } from "@/@types/utility.js"
@@ -193,9 +194,9 @@ export interface JoseInstance {
     decodeJWT: (token: string) => Promise<JWTPayload>
     encodeJWT: (payload: JWTPayload) => Promise<string>
     signJWS: (payload: JWTPayload) => Promise<string>
-    verifyJWS: (payload: string) => Promise<JWTPayload>
-    encryptJWE: (payload: string) => Promise<string>
-    decryptJWE: (payload: string) => Promise<string>
+    verifyJWS: (payload: string, options?: JWTVerifyOptions) => Promise<JWTPayload>
+    encryptJWE: (payload: string, options?: EncryptOptions) => Promise<string>
+    decryptJWE: (payload: string, options?: JWTDecryptOptions) => Promise<string>
 }
 
 export interface RouterGlobalContext {