+++ packages/core/CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Implemented timeout handling for `fetch` request used to communicate with external services during OAuth flows and user information retrieval. The implementation introduces a time limit for requests, ensuring they are canceled if they exceed the configured timeout. [#53](https://github.com/aura-stack-ts/auth/pull/53)
 
+### Changed
+
+- Update `JoseInstance` to provide full types for all parameters supported by the `verifyJWS`, `encryptJWE`, and `decryptJWE` functions exposed in the `jose` object by `createAuth` auth instance. [#55](https://github.com/aura-stack-ts/auth/pull/55)
+
 ---
 
 ## [0.2.0] - 2026-1-09